+++ lib/index.d.ts
@@ -674,6 +674,7 @@ declare class Realm {
      * @returns void
      */
     addListener(name: string, callback: (sender: Realm, event: 'change') => void): void;
+    addListener(name: string, callback: (sender: Realm, event: 'schema', schema: Realm.ObjectSchema[]) => void): void;
 
     /**
      * @param  {string} name
@@ -681,6 +682,7 @@ declare class Realm {
      * @returns void
      */
     removeListener(name: string, callback: (sender: Realm, event: 'change') => void): void;
+    removeListener(name: string, callback: (sender: Realm, event: 'schema', schema: Realm.ObjectSchema[]) => void): void;
 
     /**
      * @param  {string} name?