+++ CHANGELOG.md
@@ -7,6 +7,7 @@ x.x.x Release notes (xxxx-xx-xx)
 
 ### Enhancements
 * Added `isEmpty()` method on `Realm.Results` and `Realm.List`.
+* Added schema change listener to `Realm.addListener()` (#1825).
 
 ### Bug fixes
 * Fix `Realm.open()` to work without passing a config.