+++ CHANGELOG.md
@@ -6,6 +6,7 @@ x.x.x Release notes (xxxx-xx-xx)
 * None.
 
 ### Enhancements
+* Added `isEmpty()` method on `Realm.Results` and `Realm.List`.
 * Added schema change listener to `Realm.addListener()` (#1825).
 
 ### Bug fixes